+++ assets/javascripts/kitten/components/form/dropdown-select/combobox.js
@@ -50,7 +50,7 @@ export const DropdownCombobox = ({
   className,
   value,
   onBlur,
-  uncontrolled,
+  controlled,
 }) => {
   const [flattenedOptions, setFlattenedOptions] = useState([])
   const [filteredOptions, setFilteredOptions] = useState([])
@@ -119,7 +119,7 @@ export const DropdownCombobox = ({
     onInputValueChange,
     onIsOpenChange,
     initialIsOpen: openOnLoad,
-    ...(!uncontrolled && { selectedItem: selectedItemByValue }),
+    ...(controlled && { selectedItem: selectedItemByValue }),
   })
 
   useEffect(() => {
@@ -255,7 +255,7 @@ export const DropdownCombobox = ({
 
 DropdownCombobox.defaultProps = {
   hideLabel: false,
-  uncontrolled: false,
+  controlled: false,
   options: [],
   placeholder: 'Select',
   labelPropsGetter: () => {},
@@ -279,7 +279,7 @@ DropdownCombobox.propTypes = {
   comboboxButtonLabelText: PropTypes.string.isRequired,
   noResultText: PropTypes.string.isRequired,
   hideLabel: PropTypes.bool,
-  uncontrolled: PropTypes.bool,
+  controlled: PropTypes.bool,
   options: PropTypes.arrayOf(PropTypes.object),
   placeholder: PropTypes.string,
   labelPropsGetter: PropTypes.func,